SendDueMessageService 로직 고도화 및 DueMessageSender 이동

+++ message-scheduler/src/main/java/com/boldfaced7/fxexchange/scheduler/application/service/SendDueMessageService.java
@@ -18,7 +18,9 @@ public class SendDueMessageService implements SendDueMessageUseCase {
 
     public void sendDueScheduledMessages(SendDueMessageCommand command) {
         var messages = loadScheduledMessagePort.loadDueMessages(command.currentTimeMillis());
-        sendScheduledMessagePort.sendScheduledMessages(messages);
-        deleteScheduledMessagePort.deleteAll(messages);
+        if (!messages.isEmpty()) {
+            sendScheduledMessagePort.sendScheduledMessages(messages);
+            deleteScheduledMessagePort.deleteAll(messages);
+        }
     }
 }
